Move NoKeyOperator into used testing
