+++ nomos-services/key-management-system/src/keys/secured_key.rs
@@ -1,4 +1,4 @@
-use std::{fmt::Debug, marker::PhantomData};
+use std::fmt::Debug;
 
 use zeroize::ZeroizeOnDrop;
 
@@ -14,42 +14,6 @@ impl<T: SecureKeyOperator + Debug> DebugSecureKeyOperator for T {}
 
 pub type BoxedSecureKeyOperator<Key> =
     Box<dyn DebugSecureKeyOperator<Key = Key, Error = <Key as SecuredKey>::Error> + Send + Sync>;
-
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
-pub struct NoKeyOperator<Key, Error> {
-    _key: PhantomData<Key>,
-    _error: PhantomData<Error>,
-}
-
-#[async_trait::async_trait]
-impl<Key, Error> SecureKeyOperator for NoKeyOperator<Key, Error>
-where
-    Key: Send + Sync + 'static,
-    Error: Send + Sync + 'static,
-{
-    type Key = Key;
-    type Error = Error;
-
-    async fn execute(&mut self, _key: &Self::Key) -> Result<(), Self::Error> {
-        Ok(())
-    }
-}
-
-impl<Key, Error> Default for NoKeyOperator<Key, Error> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<Key, Error> NoKeyOperator<Key, Error> {
-    #[must_use]
-    pub const fn new() -> Self {
-        Self {
-            _key: PhantomData,
-            _error: PhantomData,
-        }
-    }
-}
 
 /// A key that can be used within the Key Management Service.
 #[async_trait::async_trait]
