+++ nomos-services/key-management-system/src/operators/mod.rs
@@ -1,1 +1,1 @@
-pub mod proofs;
+pub mod blend;
